refactor: functionsだと役割が曖昧なのでrepositoriesに移動した
