+++ src/repositories/tagSearch.ts
@@ -1,4 +1,4 @@
-import { db } from "./firebase";
+import { db } from "../repositories/firebase";
 import { collection, doc, getDocs, query, where } from "firebase/firestore";
 import { TermFragment } from "../domains/types";
