+++ src/repositories/termSetService.ts
@@ -1,4 +1,4 @@
-import { db } from "./firebase";
+import { db } from "../repositories/firebase";
 import {
   collection,
   doc,
